% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R93KarBargh2}
\alias{R93KarBargh2}
\title{Table R93KarBargh2 \cr
Income (employees in the agricultural sector) of rural households in 1393 \cr
درامد شاغلان بخش كشاورزي خانوار روستایی در سال 1393}
\format{
A data frame with 95 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{شماره ردیف فرد}
\item Description: Person row number
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{کد شغل}
\item Description: Job code
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{نوع و مشخصات فعاليت ها(زراعت)}
\item Description: Type and feature of activities (agriculture)
}
\item \code{DYCOL04}:
\itemize{
\item Farsi name: \emph{سطح زير كشت}
\item Description:
}
\item \code{DYCOL05}:
\itemize{
\item Farsi name: \emph{باغداری}
\item Description:
}
\item \code{DYCOL06}:
\itemize{
\item Farsi name: \emph{مساحت باغ}
\item Description:
}
\item \code{DYCOL07}:
\itemize{
\item Farsi name: \emph{دامداری}
\item Description:
}
\item \code{DYCOL08}:
\itemize{
\item Farsi name: \emph{تعداد دام کوچک}
\item Description:
}
\item \code{DYCOL09}:
\itemize{
\item Farsi name: \emph{تعداد دام بزرگ}
\item Description:
}
\item \code{DYCOL10}:
\itemize{
\item Farsi name: \emph{مرغداری}
\item Description:
}
\item \code{DYCOL11}:
\itemize{
\item Farsi name: \emph{تعداد مرغ، جوجه و خروس}
\item Description:
}
\item \code{DYCOL12}:
\itemize{
\item Farsi name: \emph{پرورش ماهی}
\item Description:
}
\item \code{DYCOL13}:
\itemize{
\item Farsi name: \emph{مساحت استخر}
\item Description:
}
\item \code{DYCOL14}:
\itemize{
\item Farsi name: \emph{سایر فعالیت های کشاورزی}
\item Description:
}
\item \code{DYCOL15}:
\itemize{
\item Farsi name: \emph{مزد و حقوق و مزايا (پولي و جنسي )دستمزد ميراب}
\item Description: Wages, salaries and benefits (monetary and goods). classified in DyCole16 - DyCole25
}
\item \code{DYCOL16}:
\itemize{
\item Farsi name: \emph{دستمزد دشتبان}
\item Description:
}
\item \code{DYCOL17}:
\itemize{
\item Farsi name: \emph{دستمزد دروگر}
\item Description:
}
\item \code{DYCOL18}:
\itemize{
\item Farsi name: \emph{دستمزد وجین کار}
\item Description:
}
\item \code{DYCOL19}:
\itemize{
\item Farsi name: \emph{دستمزد چوپان}
\item Description:
}
\item \code{DYCOL20}:
\itemize{
\item Farsi name: \emph{دستمزد چوبکش}
\item Description:
}
\item \code{DYCOL21}:
\itemize{
\item Farsi name: \emph{دستمزد میوه چین}
\item Description:
}
\item \code{DYCOL22}:
\itemize{
\item Farsi name: \emph{دستمزد باربر}
\item Description:
}
\item \code{DYCOL23}:
\itemize{
\item Farsi name: \emph{دستمزد باغبان}
\item Description:
}
\item \code{DYCOL24}:
\itemize{
\item Farsi name: \emph{سایر دستمزدهای پرداختی}
\item Description:
}
\item \code{DYCOL25}:
\itemize{
\item Farsi name: \emph{جمع هزینه ها}
\item Description:
}
\item \code{DYCOL26}:
\itemize{
\item Farsi name: \emph{بذر، كود، سم، علوفه، هزينه سوخت و تعميرات لوازم كار و نظاير آن- تهيه دام براي پروار بندي}
\item Description:
}
\item \code{DYCOL27}:
\itemize{
\item Farsi name: \emph{تهیه انواع بذر و نهال}
\item Description:
}
\item \code{DYCOL28}:
\itemize{
\item Farsi name: \emph{تهیه کود شیمیایی و ویتامین های مختلف}
\item Description:
}
\item \code{DYCOL29}:
\itemize{
\item Farsi name: \emph{تهیه کود حیوانی از محل اشتغال دامداری}
\item Description:
}
\item \code{DYCOL30}:
\itemize{
\item Farsi name: \emph{کود حیوانی خریداری شده}
\item Description:
}
\item \code{DYCOL31}:
\itemize{
\item Farsi name: \emph{کود حیوانی خریداری شده از محل اشتغال}
\item Description:
}
\item \code{DYCOL32}:
\itemize{
\item Farsi name: \emph{سوخت تراکتور کمباین}
\item Description:
}
\item \code{DYCOL33}:
\itemize{
\item Farsi name: \emph{سوخت تیلر وسایل و سایر وسایل نقلیه}
\item Description:
}
\item \code{DYCOL34}:
\itemize{
\item Farsi name: \emph{تعمیرات موتور آب}
\item Description:
}
\item \code{DYCOL35}:
\itemize{
\item Farsi name: \emph{تعمیرات تراکتور، کمباین}
\item Description:
}
\item \code{DYCOL36}:
\itemize{
\item Farsi name: \emph{تعمیرات خرمن کوب، تیلر}
\item Description:
}
\item \code{DYCOL37}:
\itemize{
\item Farsi name: \emph{تعمیرات سایر وسایل نقلیه}
\item Description:
}
\item \code{DYCOL38}:
\itemize{
\item Farsi name: \emph{سوخت موتور آب}
\item Description:
}
\item \code{DYCOL39}:
\itemize{
\item Farsi name: \emph{تهیه سم خریداری شده و سم پاش}
\item Description:
}
\item \code{DYCOL40}:
\itemize{
\item Farsi name: \emph{تیزکردن ابزار کار}
\item Description:
}
\item \code{DYCOL41}:
\itemize{
\item Farsi name: \emph{تهیه بنزین و گازوئیل وسایل نقلیه}
\item Description:
}
\item \code{DYCOL42}:
\itemize{
\item Farsi name: \emph{تهیه علوفه و کاه}
\item Description:
}
\item \code{DYCOL43}:
\itemize{
\item Farsi name: \emph{شکر، موم و قاب برای زنبور عسل}
\item Description:
}
\item \code{DYCOL44}:
\itemize{
\item Farsi name: \emph{سایر هزینه های مشابه}
\item Description:
}
\item \code{DYCOL45}:
\itemize{
\item Farsi name: \emph{جمع هزینه ها}
\item Description:
}
\item \code{DYCOL46}:
\itemize{
\item Farsi name: \emph{تهيه ابزار كار بي دوام و هزينه ي استهلاك ابزار و ماشين آلات بادوام – سطل}
\item Description:
}
\item \code{DYCOL47}:
\itemize{
\item Farsi name: \emph{دبه پلاستیکی}
\item Description:
}
\item \code{DYCOL48}:
\itemize{
\item Farsi name: \emph{استهلاک موتور آب}
\item Description:
}
\item \code{DYCOL49}:
\itemize{
\item Farsi name: \emph{استهلاک تیلر و خرمنکوب}
\item Description:
}
\item \code{DYCOL50}:
\itemize{
\item Farsi name: \emph{استهلاک وسایل نقلیه شغل زراعت}
\item Description:
}
\item \code{DYCOL51}:
\itemize{
\item Farsi name: \emph{استهلاک سایر وسایل عمده کار}
\item Description:
}
\item \code{DYCOL52}:
\itemize{
\item Farsi name: \emph{تعمیرات غیراساسی}
\item Description:
}
\item \code{DYCOL53}:
\itemize{
\item Farsi name: \emph{سایر}
\item Description:
}
\item \code{DYCOL54}:
\itemize{
\item Farsi name: \emph{جمع هزینه ها}
\item Description:
}
\item \code{DYCOL55}:
\itemize{
\item Farsi name: \emph{کارمزد وام شغلی، اجاره ابزار و زمین، حمل و نقل و بسته بندی و نظایر آن-کارمزد وام شغلی}
\item Description:
}
\item \code{DYCOL56}:
\itemize{
\item Farsi name: \emph{اجاره تراکتور(باراننده و بدون راننده)}
\item Description:
}
\item \code{DYCOL57}:
\itemize{
\item Farsi name: \emph{اجاره کمباین}
\item Description:
}
\item \code{DYCOL58}:
\itemize{
\item Farsi name: \emph{اجاره خرمنکوب و تیلر}
\item Description:
}
\item \code{DYCOL59}:
\itemize{
\item Farsi name: \emph{اجاره گاو کار و خیش}
\item Description:
}
\item \code{DYCOL60}:
\itemize{
\item Farsi name: \emph{اجاره زمین و باغ(نقدی و جنسی)}
\item Description:
}
\item \code{DYCOL61}:
\itemize{
\item Farsi name: \emph{حمل و نقل محصول}
\item Description:
}
\item \code{DYCOL62}:
\itemize{
\item Farsi name: \emph{تهیه گونی، جعبه، کیسه نایلونی، طناب، کارتون مقوایی و غیره}
\item Description:
}
\item \code{DYCOL63}:
\itemize{
\item Farsi name: \emph{لایروبی چاه}
\item Description:
}
\item \code{DYCOL64}:
\itemize{
\item Farsi name: \emph{آب}
\item Description:
}
\item \code{DYCOL65}:
\itemize{
\item Farsi name: \emph{برق}
\item Description:
}
\item \code{DYCOL66}:
\itemize{
\item Farsi name: \emph{سایر هزینه های مشابه}
\item Description:
}
\item \code{DYCOL67}:
\itemize{
\item Farsi name: \emph{جمع هزینه ها}
\item Description:
}
\item \code{DYCOL68}:
\itemize{
\item Farsi name: \emph{محصولات کشاورزی}
\item Description:
}
\item \code{DYCOL69}:
\itemize{
\item Farsi name: \emph{کاه و علوفه}
\item Description:
}
\item \code{DYCOL70}:
\itemize{
\item Farsi name: \emph{محصولات جالیزی}
\item Description:
}
\item \code{DYCOL71}:
\itemize{
\item Farsi name: \emph{محصولات صیفی}
\item Description:
}
\item \code{DYCOL72}:
\itemize{
\item Farsi name: \emph{سبزی های برگی}
\item Description:
}
\item \code{DYCOL73}:
\itemize{
\item Farsi name: \emph{سبزی های بوته ای}
\item Description:
}
\item \code{DYCOL74}:
\itemize{
\item Farsi name: \emph{سبزی های ریشه ای}
\item Description:
}
\item \code{DYCOL75}:
\itemize{
\item Farsi name: \emph{میوه های درختی(جز مرکبات)}
\item Description:
}
\item \code{DYCOL76}:
\itemize{
\item Farsi name: \emph{محصولات درختی(مرکبات)}
\item Description:
}
\item \code{DYCOL77}:
\itemize{
\item Farsi name: \emph{اصله و نهال}
\item Description:
}
\item \code{DYCOL78}:
\itemize{
\item Farsi name: \emph{خاک از بذر گل ها}
\item Description:
}
\item \code{DYCOL79}:
\itemize{
\item Farsi name: \emph{گل}
\item Description:
}
\item \code{DYCOL80}:
\itemize{
\item Farsi name: \emph{آب}
\item Description:
}
\item \code{DYCOL81}:
\itemize{
\item Farsi name: \emph{لبنیات}
\item Description:
}
\item \code{DYCOL82}:
\itemize{
\item Farsi name: \emph{پشم، مو، کرک و پرمرغ}
\item Description:
}
\item \code{DYCOL83}:
\itemize{
\item Farsi name: \emph{کودهای حیوانی}
\item Description:
}
\item \code{DYCOL84}:
\itemize{
\item Farsi name: \emph{انواع دام}
\item Description:
}
\item \code{DYCOL85}:
\itemize{
\item Farsi name: \emph{جوجه های تولیدی}
\item Description:
}
\item \code{DYCOL86}:
\itemize{
\item Farsi name: \emph{زنبورهای تولیدی}
\item Description:
}
\item \code{DYCOL87}:
\itemize{
\item Farsi name: \emph{عسل و موم}
\item Description:
}
\item \code{DYCOL88}:
\itemize{
\item Farsi name: \emph{تخم مرغ}
\item Description:
}
\item \code{DYCOL89}:
\itemize{
\item Farsi name: \emph{سایر اغلام مشابه}
\item Description:
}
\item \code{DYCOL90}:
\itemize{
\item Farsi name: \emph{جمع فروش}
\item Description:
}
}
}
\usage{
R93KarBargh2
}
\description{
{R for Rural} \cr
{93 for 1393}
}
\keyword{datasets}
