% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R93P3S12}
\alias{R93P3S12}
\title{Table R93P3S12 \cr
Household miscellaneous expenditure of rural household in 1393 \cr
هزينه هاي  كالاها و خدمات متفرقه خانوار روستایی سال 1393}
\format{
A data frame with 4 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس خانوار}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{طریق تهیه}
\item Description: Preparation method
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{ارزش (ریال)}
\item Description: Value (Rials)
}
}
}
\usage{
R93P3S12
}
\description{
{R for Rural} \cr
{93 for 1393} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
