% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R93P4S01}
\alias{R93P4S01}
\title{Table R93P4S01 \cr
Net wage and salary income of public, cooperative and private sectors of rural household in 1393 \cr
درآمد پولي اعضای شاغل خانوار روستایی از مشاغل مزد و حقوق بگيری سال 1393}
\format{
A data frame with 16 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{شماره ردیف عضو شاغل}
\item Description: Row number of the working member
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{آيا در حال حاضر شاغل است؟}
\item Description: Is she/he currently employed?
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{كدشغل}
\item Description: Job code
}
\item \code{DYCOL04}:
\itemize{
\item Farsi name: \emph{فعاليت اصلي محل كار}
\item Description: Main activity of the workplace
}
\item \code{DYCOL05}:
\itemize{
\item Farsi name: \emph{بخش عمومي 1 بخش تعاوني 2 بخش خصوصي 3}
\item Description: General Section 1 Cooperative Section 2 Private Section 3
}
\item \code{DYCOL06}:
\itemize{
\item Farsi name: \emph{ميزان ساعت هاي كار در روز}
\item Description: The amount of working hours per day
}
\item \code{DYCOL07}:
\itemize{
\item Farsi name: \emph{تعداد روزهاي كار در هفته}
\item Description: Number of working days per week
}
\item \code{DYCOL08}:
\itemize{
\item Farsi name: \emph{مجموع درآمدهای ناخالص مستمر و غیرمستمر قبل از کسورات(ماه گذشته)}
\item Description: Total of continuous and non-continuous gross income before deficits(last month)
}
\item \code{DYCOL09}:
\itemize{
\item Farsi name: \emph{مجموع درآمدهای ناخالص مستمر و غیرمستمر قبل از کسورات(12 ماه گذشته)}
\item Description: Total of continuous and non-continuous gross income before deficits(last 12 months)
}
\item \code{DYCOL10}:
\itemize{
\item Farsi name: \emph{مزد و حقوق و مزایای مستمر (ماه گذشته)}
\item Description: Continuous wages and benefits (last Month)
}
\item \code{DYCOL11}:
\itemize{
\item Farsi name: \emph{مزد و حقوق و مزایای مستر (12 ماه گذشته)}
\item Description: Continuous wages and benefits (last 12 Months)
}
\item \code{DYCOL12}:
\itemize{
\item Farsi name: \emph{مزایای غیرمستمر (ماه گذشته)}
\item Description: Noncontinuous benefits (last month)
}
\item \code{DYCOL13}:
\itemize{
\item Farsi name: \emph{مزایای غیرمستمر (12 ماه گذشته)}
\item Description: Noncontinuous benefits (last 12 months)TV
}
\item \code{DYCOL14}:
\itemize{
\item Farsi name: \emph{مجموع درآمد خالص(ماه گذشته)}
\item Description: Total net income (last month)
}
\item \code{DYCOL15}:
\itemize{
\item Farsi name: \emph{مجموع درآمد خالص (12 ماه گذشته)}
\item Description: Total net income (last 12 months)
}
}
}
\usage{
R93P4S01
}
\description{
{R for Rural} \cr
{93 for 1393} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
