% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U93KarBargh1}
\alias{U93KarBargh1}
\title{Table U93KarBargh1 \cr
Worksheet related to miscellaneous incomes of urban household members in 1393 \cr
كاربرگ مربوط به درامدهاي متفرقه اعضاء خانوار شهری سال 1393}
\format{
A data frame with 4 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{شماره ردیف فرد}
\item Description: Person row number
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{شماره ردیف فعالیت}
\item Description: Activity row number
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{درآمد}
\item Description: Income
}
}
}
\usage{
U93KarBargh1
}
\description{
{U for Urban} \cr
{93 for 1393}
}
\keyword{datasets}
