% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U93P3S01}
\alias{U93P3S01}
\title{Table U93P3S01 \cr
Household food and tobacco expenditure of urban household in 1393 \cr
هزينه هاي خوراكي و دخانی خانوار شهری سال 1393}
\format{
A data frame with 7 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{طریق تهیه}
\item Description: Preparation method
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{مقدار (گرم)}
\item Description: Amount(gram)
}
\item \code{DYCOL04}:
\itemize{
\item Farsi name: \emph{مقدار (کیلوگرم)}
\item Description: Amount(kilogram)
}
\item \code{DYCOL05}:
\itemize{
\item Farsi name: \emph{قیمت واحد (ریال)}
\item Description: Unit price (Rials)
}
\item \code{DYCOL06}:
\itemize{
\item Farsi name: \emph{ارزش (ریال)}
\item Description: Value (Rials)
}
}
}
\usage{
U93P3S01
}
\description{
{U for Urban} \cr
{93 for 1393} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
