% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U93P3S04}
\alias{U93P3S04}
\title{Table U93P3S04 \cr
Current household housing expenditure of urban household in 1393 \cr
هزينه هاي جاری مسكن خانوار شهری سال 1393}
\format{
A data frame with 5 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس خانوار}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{مبلغ رهن}
\item Description: Mortgage amount
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{طریق تهیه}
\item Description: Preparation method
}
\item \code{DYCOL04}:
\itemize{
\item Farsi name: \emph{ارزش (ریال)}
\item Description: Value (Rials)
}
}
}
\usage{
U93P3S04
}
\description{
{U for Urban} \cr
{93 for 1393} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
