% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U93P4S02}
\alias{U93P4S02}
\title{Table U93P4S02 \cr
Self-employment income of urban household in 1393 \cr
درآمد پولي اعضای شاغل خانوار شهری از مشاغل آزاد سال 1393}
\format{
A data frame with 16 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{شماره ردیف عضو شاغل}
\item Description: Row number of the working member
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{آيا در حال حاضر شاغل است؟}
\item Description: Is she/he currently employed?
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{كدشغل}
\item Description: Job code
}
\item \code{DYCOL04}:
\itemize{
\item Farsi name: \emph{فعاليت اصلي محل كار}
\item Description: Main activity of the workplace
}
\item \code{DYCOL05}:
\itemize{
\item Farsi name: \emph{کارفرما 1 کارکن مستقل 2 کارکن فامیلی 3}
\item Description: Employer 1 independent employee 2 family employee 3
}
\item \code{DYCOL06}:
\itemize{
\item Farsi name: \emph{كد شغل کشاورزی 1 غیرکشاورزی 2}
\item Description: Agriculture 1 non-agriculture 2
}
\item \code{DYCOL07}:
\itemize{
\item Farsi name: \emph{ميزان ساعت هاي كار در روز}
\item Description: The amount of working hours per day
}
\item \code{DYCOL08}:
\itemize{
\item Farsi name: \emph{تعداد روزهاي كار در هفته}
\item Description: Number of working days per week
}
\item \code{DYCOL09}:
\itemize{
\item Farsi name: \emph{مزد و حقوق و مزایا}
\item Description: Wages and benefits
}
\item \code{DYCOL10}:
\itemize{
\item Farsi name: \emph{بذر، آب و کود}
\item Description: Seeds, water and fertilizer
}
\item \code{DYCOL11}:
\itemize{
\item Farsi name: \emph{تهیه ابزار کار بی دوام}
\item Description: Preparation of unsustainable working tools
}
\item \code{DYCOL12}:
\itemize{
\item Farsi name: \emph{کارمزد شغلی}
\item Description: Job wage
}
\item \code{DYCOL13}:
\itemize{
\item Farsi name: \emph{مالیات شغلی}
\item Description: Job tax
}
\item \code{DYCOL14}:
\itemize{
\item Farsi name: \emph{فروش (دریافتی ناخالص)}
\item Description: Sales (gross receiving)
}
\item \code{DYCOL15}:
\itemize{
\item Farsi name: \emph{درآمد خالص}
\item Description: Net income
}
}
}
\usage{
U93P4S02
}
\description{
{U for Urban} \cr
{93 for 1393} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
