% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U93P4S03}
\alias{U93P4S03}
\title{Table U93P4S03 \cr
Miscellaneous income of urban household in 1393 \cr
درآمد هاي متفرقه خانوار شهری سال 1393}
\format{
A data frame with 8 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{شماره ردیف اعضای خانوار}
\item Description: Row number of family members
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{حقوق بازنشستگی}
\item Description: Retirement wage
}
\item \code{DYCOL04}:
\itemize{
\item Farsi name: \emph{درآمد حاصل از اجاره محل کسب}
\item Description: Revenue from renting the business place
}
\item \code{DYCOL05}:
\itemize{
\item Farsi name: \emph{درآمد حاصل از حساب پس انداز}
\item Description: Income from the savings account
}
\item \code{DYCOL06}:
\itemize{
\item Farsi name: \emph{کمک هزینه تحصیلی}
\item Description: Scholarships
}
\item \code{DYCOL07}:
\itemize{
\item Farsi name: \emph{درآمد حاصل از محل فروش مصنوعات}
\item Description: Income from the sale of artifacts
}
\item \code{DYCOL08}:
\itemize{
\item Farsi name: \emph{دریافتی های انتقالی از خانوارهای دیگر}
\item Description: Transfer receivables from other families
}
}
}
\usage{
U93P4S03
}
\description{
{U for Urban} \cr
{93 for 1393} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
